update tests after merging with 2.0.0
