+++ collapsevariants/collapse_utils.py
@@ -4,33 +4,12 @@ from typing import Tuple, Dict, List
 import numpy as np
 import pandas as pd
 from bgen import BgenReader
-from general_utilities.association_resources import download_dxfile_by_name
 from general_utilities.mrc_logger import MRCLogger
 from scipy.sparse import coo_matrix, csr_matrix
 
 from collapsevariants.collapse_logger import CollapseLOGGER
-from collapsevariants.ingest_data import BGENIndex
 
 LOGGER = MRCLogger().get_logger()
-
-
-def download_bgen(chrom_bgen_index: BGENIndex) -> Tuple[Path, Path, Path]:
-    """Download the BGEN file from DNANexus
-
-    This method downloads the BGEN file from DNANexus using the bgen_index dictionary and the bgen_prefix. It then
-    returns the path to the downloaded BGEN file, index, and sample.
-
-    Note that this method is the only method untested in this package as it requires a DNANexus connection to test.
-
-    :return: A Tuple containing the paths to the BGEN file, index file, and sample file
-    """
-
-    # Download the requisite files for this chromosome according to the index dict:
-    bgen_path = download_dxfile_by_name(chrom_bgen_index['bgen'], print_status=False)
-    index_path = download_dxfile_by_name(chrom_bgen_index['index'], print_status=False)
-    sample_path = download_dxfile_by_name(chrom_bgen_index['sample'], print_status=False)
-
-    return bgen_path, index_path, sample_path
 
 
 def get_sample_ids(sample_path: Path) -> List[str]:
